Refactored options
