+++ src/options.js
@@ -1,0 +1,6 @@
+export const buildNullInputParams = (filter, json) => {
+  return [
+    '--null-input',
+    `${JSON.stringify(json)} | ${filter}`
+  ]
+}
