+++ test/jq.test.js
@@ -3,9 +3,7 @@ import chaiAsPromised from 'chai-as-promised'
 chai.use(chaiAsPromised)
 import { run } from '../src/jq'
 import path from 'path'
-import jsonFixture from './fixtures/1.json'
 
-const jsonFixtureToString = JSON.stringify(jsonFixture)
 const FIXTURES_PATH = path.join(__dirname, 'fixtures')
 const ROOT_PATH = path.join(__dirname, '..')
 const jsonPathFixture = path.join(FIXTURES_PATH, '1.json')
@@ -13,7 +11,7 @@ const jsPathFixture = path.join(FIXTURES_PATH, '1.js')
 const asteriskPathFixture = path.join(ROOT_PATH, 'src', '*.js')
 const filter = '. | map(select(.a == .id))'
 
-describe('jq runs a cli', () => {
+describe('jq core', () => {
   it('should return a stdout object', () => {
     return expect(
       run(filter, jsonPathFixture)
@@ -36,11 +34,5 @@ describe('jq runs a cli', () => {
     return expect(
       run(filter, jsPathFixture)
     ).to.eventually.be.rejectedWith(Error)
-  })
-
-  it('should run the filter inline with null-input', () => {
-    return expect(
-      run('.', jsonFixtureToString, { nullInput: true })
-    ).to.eventually.become(jsonFixtureToString)
   })
 })
